add new tab for report
