+++ web-ui/src/main/resources/catalog/locales/en-admin.json
@@ -396,6 +396,7 @@
     "harvesterRecords": "Harvested records",
     "harvesterSchedule": "Schedule",
     "harvesterScheduleinactive": "Failed to activate harvester.",
+    "harvesterType": "Node type",
     "harvesterUpdated": "Harvester updated",
     "harvesterValidate": "Validate records before import",
     "harvesterValidateHelp": "Invalid records will be rejected. Validation is based on the standard validation (ie. XSD, Schematrons).",
@@ -590,6 +591,7 @@
     "reportFileUploadMetadata-title": "Metadata file uploads report",
     "reportFileUploadMetadata-uploadDateBegin": "Upload date (begin)",
     "reportFileUploadMetadata-uploadDateEnd": "Upload date (end)",
+    "reportHarvester" : "Harvesters report",
     "reportInternalMetadata": "Internal metadata",
     "reportInternalMetadata-description": "Select criteria to extract a list of internal records and the record owner’s information. This will allow for a record review to ensure completeness of records or if necessary the deletion of “blank” or “test” records. The report results can be imported into MS Access and Excel.",
     "reportInternalMetadata-metadataCreationDateBegin": "Metadata creation date (begin)",
