+++ web-ui/src/main/resources/catalog/locales/fr-admin.json
@@ -389,6 +389,7 @@
     "harvesterRecords": "Fiches moissonnées",
     "harvesterSchedule": "Planification",
     "harvesterScheduleinactive": "Échec de l'activation du noeud.",
+    "harvesterType": "Type de noeud",
     "harvesterUpdated": "Noeud mis à jour",
     "harvesterValidate": "Valider les fiches avant l'import",
     "harvesterValidateHelp": "Les fiches invalides sont alors rejetées. La validation repose sur le standards (ie. XSD, Schematrons).",
@@ -584,6 +585,7 @@
     "reportFileUploadMetadata-title": "Rapport sur l'ajout de fichiers aux fiches",
     "reportFileUploadMetadata-uploadDateBegin": "Depuis",
     "reportFileUploadMetadata-uploadDateEnd": "à",
+    "reportHarvester" : "Rapport des noeuds",
     "reportInternalMetadata": "Fiches internes",
     "reportInternalMetadata-description": "Ce rapport permet d'extraire la liste des fiches internes. This will allow for a record review to ensure completeness of records or if necessary the deletion of “blank” or “test” records. Ce rapport est au format CSV.",
     "reportInternalMetadata-metadataCreationDateBegin": "Depuis",
@@ -873,12 +875,20 @@
     "thredds-templateIdHelp": " ",
     "thredds-topic": "Thème",
     "thredds-unidataCfg": "UNIDATA configuration",
+
+    "metadataschemaValidate" : "Validation",
+    "metadataschemaValidateDes" : "Configure what rules are applied for each schema metadata. If no rule defined for a schema, it will be applied always.",
     "schematroncriteriagroup" : "Groupe de critères",
     "xpathlabel" : "xPath",
     "xpathvalue" : "Value",
     "toggleRequirement" : "Change Requirement",
     "deleteCriteriaGroup" : "Supprimer le groupe de critères",
     "deleteCriteria" : "Supprimer le critère",
+    "harvester": "Noeuds distants",
+    "report":"Rapport",
+    "newHarvester": "Moissonner ",
+    "deleteHarvester": "Supprimer",
+    "deleteHarvesterHelp": "Supprimer le noeud et tous les fiches associées",
     "recordsRemoved": "fiche(s) supprimée(s)",
     "time": "Temps",
     "tools": "Outils",
