+++ web-ui/src/main/resources/catalog/style/gn.less
@@ -93,6 +93,9 @@
 a {
   word-wrap: break-word;
 }
+.gn-break {
+  word-wrap: break-word;
+}
 
 
 /**
